+++ kbe/src/lib/navigation/navigation_tile_handle.cpp
@@ -432,7 +432,6 @@ int NavTileHandle::getMap(int x, int y)
 //-------------------------------------------------------------------------------------
 bool NavTileHandle::MapSearchNode::IsSameState(MapSearchNode &rhs)
 {
-
 	// same state in a maze search is simply when (x,y) are the same
 	if( (x == rhs.x) &&
 		(y == rhs.y) )