+++ kbe/src/lib/server/pendingLoginmgr.h
@@ -54,6 +54,7 @@ public:
 			entityID = 0;
 			lastProcessTime = 0;
 			forceInternalLogin = false;
+			needCheckPassword = true;
 		}
 
 		Network::Address addr;
@@ -67,6 +68,7 @@ public:
 		uint32 flags;
 		uint64 deadline;
 		bool forceInternalLogin;
+		bool needCheckPassword;
 	};
 
 	typedef KBEUnordered_map<std::string, PLInfos*> PTINFO_MAP;