+++ kbe/src/lib/server/telnet_server.cpp
@@ -23,6 +23,7 @@ along with KBEngine.  If not, see <http://www.gnu.org/licenses/>.
 #include "telnet_handler.h"
 #include "network/bundle.h"
 #include "network/endpoint.h"
+#include "network/network_interface.h"
 
 #ifndef CODE_INLINE
 #include "telnet_server.inl"
@@ -171,8 +172,8 @@ int	TelnetServer::handleInputNotification(int fd)
 
 			if(tickcount == 1)
 			{
-				WARNING_MSG(fmt::format("TelnetServer::handleInputNotification: accept endpoint({}) {}!\n",
-					 fd, kbe_strerror()));
+				WARNING_MSG(fmt::format("TelnetServer::handleInputNotification: accept endpoint({}) {}! channelSize={}\n",
+					fd, kbe_strerror(), pNetworkInterface_->channels().size()));
 			}
 
 			break;