+++ kbe/src/server/baseapp/baseapp.cpp
@@ -187,7 +187,8 @@ Baseapp::Baseapp(Network::EventDispatcher& dispatcher,
 	pRestoreEntityHandlers_(),
 	pResmgrTimerHandle_(),
 	pInitProgressHandler_(NULL),
-	flags_(APP_FLAGS_NONE)
+	flags_(APP_FLAGS_NONE),
+	pBundleImportEntityDefDatas_(NULL)
 {
 	KBEngine::Network::MessageHandlers::pMainMessageHandlers = &BaseappInterface::messageHandlers;
 
@@ -590,6 +591,12 @@ void Baseapp::finalise()
 	loopCheckTimerHandle_.cancel();
 	pResmgrTimerHandle_.cancel();
 	forward_messagebuffer_.clear();
+
+	if (pBundleImportEntityDefDatas_)
+	{
+		Network::Bundle::reclaimPoolObject(pBundleImportEntityDefDatas_);
+		pBundleImportEntityDefDatas_ = NULL;
+	}
 
 	EntityApp<Base>::finalise();
 }
@@ -3162,9 +3169,11 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 	uint32 nrows = 0;
 	uint32 nfields = 0;
 	uint64 affectedRows = 0;
+	uint64 lastInsertID = 0;
 
 	PyObject* pResultSet = NULL;
 	PyObject* pAffectedRows = NULL;
+	PyObject* pLastInsertID = NULL;
 	PyObject* pErrorMsg = NULL;
 
 	s >> callbackID;
@@ -3181,6 +3190,9 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 		{
 			pAffectedRows = Py_None;
 			Py_INCREF(pAffectedRows);
+
+			pLastInsertID = Py_None;
+			Py_INCREF(pLastInsertID);
 
 			s >> nrows;
 
@@ -3222,6 +3234,9 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 			s >> affectedRows;
 
 			pAffectedRows = PyLong_FromUnsignedLongLong(affectedRows);
+
+			s >> lastInsertID;
+			pLastInsertID = PyLong_FromUnsignedLongLong(lastInsertID);
 		}
 	}
 	else
@@ -3233,6 +3248,9 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 
 			pAffectedRows = Py_None;
 			Py_INCREF(pAffectedRows);
+
+			pLastInsertID = Py_None;
+			Py_INCREF(pLastInsertID);
 	}
 
 	s.done();
@@ -3248,8 +3266,8 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 		if(pyfunc != NULL)
 		{
 			PyObject* pyResult = PyObject_CallFunction(pyfunc.get(), 
-												const_cast<char*>("OOO"), 
-												pResultSet, pAffectedRows, pErrorMsg);
+												const_cast<char*>("OOOO"), 
+												pResultSet, pAffectedRows, pLastInsertID, pErrorMsg);
 
 			if(pyResult != NULL)
 				Py_DECREF(pyResult);
@@ -3265,6 +3283,7 @@ void Baseapp::onExecuteRawDatabaseCommandCB(Network::Channel* pChannel, KBEngine
 
 	Py_XDECREF(pResultSet);
 	Py_XDECREF(pAffectedRows);
+	Py_XDECREF(pLastInsertID);
 	Py_XDECREF(pErrorMsg);
 }
 
@@ -3567,8 +3586,9 @@ void Baseapp::registerPendingLogin(Network::Channel* pChannel, KBEngine::MemoryS
 	uint64										deadline;
 	COMPONENT_TYPE								componentType;
 	bool										forceInternalLogin;
+	bool										needCheckPassword;
 
-	s >> loginName >> accountName >> password >> entityID >> entityDBID >> flags >> deadline >> componentType >> forceInternalLogin;
+	s >> loginName >> accountName >> password >> needCheckPassword >> entityID >> entityDBID >> flags >> deadline >> componentType >> forceInternalLogin;
 	s.readBlob(datas);
 
 	Network::Bundle* pBundle = Network::Bundle::createPoolObject();
@@ -3598,6 +3618,7 @@ void Baseapp::registerPendingLogin(Network::Channel* pChannel, KBEngine::MemoryS
 	ptinfos->deadline = deadline;
 	ptinfos->ctype = (COMPONENT_CLIENT_TYPE)componentType;
 	ptinfos->datas = datas;
+	ptinfos->needCheckPassword = needCheckPassword;
 	pendingLoginMgr_.add(ptinfos);
 }
 
@@ -3714,6 +3735,15 @@ void Baseapp::loginBaseapp(Network::Channel* pChannel,
 		return;
 	}
 
+	// 虽然接入第三方dbmgr不检查密码，但至少在loginapp时提交的password应该跟本次提交的能匹配上
+	// 否则容易被其他连接攻击式的试探登陆
+	if (!ptinfos->needCheckPassword && ptinfos->password != password)
+	{
+		loginBaseappFailed(pChannel, accountName, SERVER_ERR_NAME_PASSWORD);
+		pendingLoginMgr_.removeNextTick(accountName);
+		return;
+	}
+
 	// 如果entityID大于0则说明此entity是存活状态登录
 	if(ptinfos->entityID > 0)
 	{
@@ -3805,7 +3835,7 @@ void Baseapp::loginBaseapp(Network::Channel* pChannel,
 		ENTITY_ID entityID = idClient_.alloc();
 		KBE_ASSERT(entityID > 0);
 
-		DbmgrInterface::queryAccountArgs7::staticAddToBundle((*pBundle), accountName, password, g_componentID, 
+		DbmgrInterface::queryAccountArgs8::staticAddToBundle((*pBundle), accountName, password, ptinfos->needCheckPassword, g_componentID,
 			entityID, ptinfos->entityDBID, pChannel->addr().ip, pChannel->addr().port);
 
 		dbmgrinfos->pChannel->send(pBundle);
@@ -4360,7 +4390,14 @@ void Baseapp::onRemoteCallCellMethodFromClient(Network::Channel* pChannel, KBEng
 
 	ENTITY_ID srcEntityID = pChannel->proxyID();
 	if(srcEntityID <= 0)
+	{
+		ERROR_MSG(fmt::format("Baseapp::onRemoteCallCellMethodFromClient: pChannel does not bind proxy! addr={}\n",
+			pChannel->c_str()));
+				
+		pChannel->condemn();
+		s.done();
 		return;
+	}
 	
 	if(s.length() <= 0)
 		return;
@@ -4400,6 +4437,10 @@ void Baseapp::onUpdateDataFromClient(Network::Channel* pChannel, KBEngine::Memor
 	ENTITY_ID srcEntityID = pChannel->proxyID();
 	if(srcEntityID <= 0)
 	{
+		ERROR_MSG(fmt::format("Baseapp::onUpdateDataFromClient: pChannel does not bind proxy! addr={}\n",
+			pChannel->c_str()));
+				
+		pChannel->condemn();
 		s.done();
 		return;
 	}
@@ -4702,10 +4743,10 @@ void Baseapp::importClientMessages(Network::Channel* pChannel)
 //-------------------------------------------------------------------------------------
 void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 {
-	static Network::Bundle bundle;
-	
-	if(bundle.empty())
+	if (!pBundleImportEntityDefDatas_)
 	{
+		pBundleImportEntityDefDatas_ = Network::Bundle::createPoolObject();
+
 		ENTITY_PROPERTY_UID posuid = ENTITY_BASE_PROPERTY_UTYPE_POSITION_XYZ;
 		ENTITY_PROPERTY_UID diruid = ENTITY_BASE_PROPERTY_UTYPE_DIRECTION_ROLL_PITCH_YAW;
 		ENTITY_PROPERTY_UID spaceuid = ENTITY_BASE_PROPERTY_UTYPE_SPACEID;
@@ -4724,20 +4765,20 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 		if(msgInfo != NULL)
 			spaceuid = msgInfo->msgid;
 
-		bundle.newMessage(ClientInterface::onImportClientEntityDef);
+		pBundleImportEntityDefDatas_->newMessage(ClientInterface::onImportClientEntityDef);
 		
 		const DataTypes::UID_DATATYPE_MAP& dataTypes = DataTypes::uid_dataTypes();
 		uint16 aliassize = (uint16)dataTypes.size();
-		bundle << aliassize;
+		(*pBundleImportEntityDefDatas_) << aliassize;
 
 		DataTypes::UID_DATATYPE_MAP::const_iterator dtiter = dataTypes.begin();
 		for(; dtiter != dataTypes.end(); ++dtiter)
 		{
 			const DataType* datatype = dtiter->second;
 
-			bundle << datatype->id();
-			bundle << datatype->getName();
-			bundle << datatype->aliasName();
+			(*pBundleImportEntityDefDatas_) << datatype->id();
+			(*pBundleImportEntityDefDatas_) << datatype->getName();
+			(*pBundleImportEntityDefDatas_) << datatype->aliasName();
 
 			if(strcmp(datatype->getName(), "FIXED_DICT") == 0)
 			{
@@ -4746,19 +4787,19 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 				FixedDictType::FIXEDDICT_KEYTYPE_MAP& keys = dictdatatype->getKeyTypes();
 
 				uint8 keysize = (uint8)keys.size();
-				bundle << keysize;
-				bundle << dictdatatype->moduleName();
+				(*pBundleImportEntityDefDatas_) << keysize;
+				(*pBundleImportEntityDefDatas_) << dictdatatype->moduleName();
 
 				FixedDictType::FIXEDDICT_KEYTYPE_MAP::const_iterator keyiter = keys.begin();
 				for(; keyiter != keys.end(); ++keyiter)
 				{
-					bundle << keyiter->first;
-					bundle << keyiter->second->dataType->id();
+					(*pBundleImportEntityDefDatas_) << keyiter->first;
+					(*pBundleImportEntityDefDatas_) << keyiter->second->dataType->id();
 				}
 			}
 			else if(strcmp(datatype->getName(), "ARRAY") == 0)
 			{
-				bundle << const_cast<FixedArrayType*>(static_cast<const FixedArrayType*>(datatype))->getDataType()->id();
+				(*pBundleImportEntityDefDatas_) << const_cast<FixedArrayType*>(static_cast<const FixedArrayType*>(datatype))->getDataType()->id();
 			}
 		}
 
@@ -4779,16 +4820,16 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 			uint16 size2 = (uint16)methods1.size();
 			uint16 size3 = (uint16)methods2.size();
 
-			bundle << iter->get()->getName() << iter->get()->getUType() << size << size1 << size2 << size3;
+			(*pBundleImportEntityDefDatas_) << iter->get()->getName() << iter->get()->getUType() << size << size1 << size2 << size3;
 			
 			int16 aliasID = ENTITY_BASE_PROPERTY_ALIASID_POSITION_XYZ;
-			bundle << posuid << ((uint32)ED_FLAG_ALL_CLIENTS) << aliasID << "position" << "" << DataTypes::getDataType("VECTOR3")->id();
+			(*pBundleImportEntityDefDatas_) << posuid << ((uint32)ED_FLAG_ALL_CLIENTS) << aliasID << "position" << "" << DataTypes::getDataType("VECTOR3")->id();
 
 			aliasID = ENTITY_BASE_PROPERTY_ALIASID_DIRECTION_ROLL_PITCH_YAW;
-			bundle << diruid << ((uint32)ED_FLAG_ALL_CLIENTS) << aliasID << "direction" << "" << DataTypes::getDataType("VECTOR3")->id();
+			(*pBundleImportEntityDefDatas_) << diruid << ((uint32)ED_FLAG_ALL_CLIENTS) << aliasID << "direction" << "" << DataTypes::getDataType("VECTOR3")->id();
 
 			aliasID = ENTITY_BASE_PROPERTY_ALIASID_SPACEID;
-			bundle << spaceuid << ((uint32)ED_FLAG_CELL_PRIVATE) << aliasID << "spaceID" << "" << DataTypes::getDataType("UINT32")->id();
+			(*pBundleImportEntityDefDatas_) << spaceuid << ((uint32)ED_FLAG_CELL_PRIVATE) << aliasID << "spaceID" << "" << DataTypes::getDataType("UINT32")->id();
 
 			ScriptDefModule::PROPERTYDESCRIPTION_MAP::const_iterator piter = propers.begin();
 			for(; piter != propers.end(); ++piter)
@@ -4798,7 +4839,7 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 				std::string	name = piter->second->getName();
 				std::string	defaultValStr = piter->second->getDefaultValStr();
 				uint32 flags = piter->second->getFlags();
-				bundle << properUtype << flags << aliasID << name << defaultValStr << piter->second->getDataType()->id();
+				(*pBundleImportEntityDefDatas_) << properUtype << flags << aliasID << name << defaultValStr << piter->second->getDataType()->id();
 			}
 			
 			ScriptDefModule::METHODDESCRIPTION_MAP::const_iterator miter = methods.begin();
@@ -4812,12 +4853,12 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 				const std::vector<DataType*>& args = miter->second->getArgTypes();
 				uint8 argssize = (uint8)args.size();
 
-				bundle << methodUtype << aliasID << name << argssize;
+				(*pBundleImportEntityDefDatas_) << methodUtype << aliasID << name << argssize;
 				
 				std::vector<DataType*>::const_iterator argiter = args.begin();
 				for(; argiter != args.end(); ++argiter)
 				{
-					bundle << (*argiter)->id();
+					(*pBundleImportEntityDefDatas_) << (*argiter)->id();
 				}
 			}
 
@@ -4832,12 +4873,12 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 				const std::vector<DataType*>& args = miter->second->getArgTypes();
 				uint8 argssize = (uint8)args.size();
 
-				bundle << methodUtype << aliasID << name << argssize;
+				(*pBundleImportEntityDefDatas_) << methodUtype << aliasID << name << argssize;
 				
 				std::vector<DataType*>::const_iterator argiter = args.begin();
 				for(; argiter != args.end(); ++argiter)
 				{
-					bundle << (*argiter)->id();
+					(*pBundleImportEntityDefDatas_) << (*argiter)->id();
 				}
 			}
 
@@ -4852,18 +4893,18 @@ void Baseapp::importClientEntityDef(Network::Channel* pChannel)
 				const std::vector<DataType*>& args = miter->second->getArgTypes();
 				uint8 argssize = (uint8)args.size();
 
-				bundle << methodUtype << aliasID << name << argssize;
+				(*pBundleImportEntityDefDatas_) << methodUtype << aliasID << name << argssize;
 				
 				std::vector<DataType*>::const_iterator argiter = args.begin();
 				for(; argiter != args.end(); ++argiter)
 				{
-					bundle << (*argiter)->id();
+					(*pBundleImportEntityDefDatas_) << (*argiter)->id();
 				}
 			}
 		}
 	}
 
-	pChannel->send(new Network::Bundle(bundle));
+	pChannel->send(new Network::Bundle((*pBundleImportEntityDefDatas_)));
 }
 
 //-------------------------------------------------------------------------------------
@@ -4888,6 +4929,12 @@ PyObject* Baseapp::__py_reloadScript(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 void Baseapp::reloadScript(bool fullReload)
 {
+	if (pBundleImportEntityDefDatas_)
+	{
+		Network::Bundle::reclaimPoolObject(pBundleImportEntityDefDatas_);
+		pBundleImportEntityDefDatas_ = NULL;
+	}
+
 	EntityApp<Base>::reloadScript(fullReload);
 }