+++ kbe/src/server/cellapp/entity.cpp
@@ -169,7 +169,7 @@ Entity::~Entity()
 	KBE_ASSERT(pWitness_ == NULL);
 
 	SAFE_RELEASE(pControllers_);
-	SAFE_RELEASE(pEntityCoordinateNode_);
+	KBE_ASSERT(pEntityCoordinateNode_ == NULL);
 
 	Py_DECREF(pPyPosition_);
 	pPyPosition_ = NULL;
@@ -262,8 +262,9 @@ void Entity::onDestroy(bool callScript)
 		ERROR_MSG(fmt::format("{}::onDestroy(): id={}, witnesses_count({}/{}) != 0, isReal={}, spaceID={}, position=({},{},{})\n", 
 			scriptName(), id(), witnesses_count_, witnesses_.size(), isReal(), this->spaceID(), position().x, position().y, position().z));
 
-		std::list<ENTITY_ID>::iterator it = witnesses_.begin();
-		for (; it != witnesses_.end(); ++it)
+		std::list<ENTITY_ID> witnesses_copy = witnesses_;
+		std::list<ENTITY_ID>::iterator it = witnesses_copy.begin();
+		for (; it != witnesses_copy.end(); ++it)
 		{
 			Entity *ent = Cellapp::getSingleton().findEntity((*it));
 
@@ -306,6 +307,8 @@ void Entity::onDestroy(bool callScript)
 
 	pPyPosition_->onLoseRef();
 	pPyDirection_->onLoseRef();
+
+	SAFE_RELEASE(pEntityCoordinateNode_);
 }
 
 //-------------------------------------------------------------------------------------
@@ -432,7 +435,7 @@ int Entity::pySetControlledBy(PyObject *value)
 
 	EntityMailbox* mailbox = NULL;
 
-	if (value != Py_None )
+	if (value != Py_None)
 	{
 		if (!PyObject_TypeCheck(value, EntityMailbox::getScriptType()) || !((EntityMailbox *)value)->isBase())
 		{
@@ -538,6 +541,7 @@ bool Entity::setControlledBy(EntityMailbox* controllerBaseMailbox)
 			sendControlledByStatusMessage(controllerBaseMailbox, 1);
 		}
 	}
+
 	return true;
 }
 
@@ -552,6 +556,8 @@ void Entity::sendControlledByStatusMessage(EntityMailbox* baseMailbox, int8 isCo
 	{
 		pChannel = (static_cast<EntityMailbox*>(clientMB))->getChannel();
 	}
+
+	Py_DECREF(clientMB);
 
 	if (!pChannel)
 		return;